Add "overwatch", replaces kickHandler. Work on data player loader.
